requirements.txt
PyYAML~=6.0
semver~=3.0
requests~=2.31
Jinja2~=3.1.2
